+++ react/components/WordpressProductSearchResult.tsx
@@ -47,6 +47,7 @@ const WordpressSearchResult: StorefrontFunctionComponent<Props> = ({
   const intl = useIntl()
   const [page, setPage] = useState(1)
   const [perPage, setPerPage] = useState(postsPerPage)
+  const [selectedOption, setSelectedOption] = useState(postsPerPage)
   const handles = useCssHandles(CSS_HANDLES)
   const { data: dataS } = useQuery(Settings)
   const { loading, error, data, fetchMore } = useQuery(SearchPosts, {
@@ -87,6 +88,7 @@ const WordpressSearchResult: StorefrontFunctionComponent<Props> = ({
         postsPerPage * 3,
         postsPerPage * 4,
       ]}
+      selectedOption={selectedOption}
       currentItemFrom={(page - 1) * perPage + 1}
       currentItemTo={page * perPage}
       textOf="of"
@@ -98,6 +100,7 @@ const WordpressSearchResult: StorefrontFunctionComponent<Props> = ({
       }
       totalItems={data?.wpPosts?.total_count ?? 0}
       onRowsChange={({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
+        setSelectedOption(+value)
         setPage(1)
         setPerPage(+value)
         fetchMore({